Code cleanup and structure update
