+++ AWSAppSyncClient/AWSAppSyncClientConflictResolutionExtensions.swift
@@ -40,7 +40,7 @@ extension AWSAppSyncClient {
             }
             
             firstly {
-                try response.parseResult(cacheKeyForObject: self.store!.cacheKeyForObject)
+                try response.parseResult(cacheKeyForObject: self.store.cacheKeyForObject)
                 }.andThen { (result, records) in
                     if let resultError = result.errors,
                         let conflictResolutionBlock = conflictResolutionBlock,
@@ -61,7 +61,7 @@ extension AWSAppSyncClient {
                     } else {
                         notifyResultHandler(result: result, error: nil)
                         if let records = records {
-                            self.store?.publish(records: records, context: context).catch { error in
+                            self.store.publish(records: records, context: context).catch { error in
                                 preconditionFailure(String(describing: error))
                             }
                         }
